pyproject.toml
[project]
name = "vocachieve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
